+++ retry/retry.go
@@ -46,3 +46,14 @@ func OnError(maxTries int, retriable func(error) bool, workload func() error) er
 	}
 	return err
 }
+
+// OnConflict provides a K8s-way "retrier" mechanism to avoid conflicts on resource updates.
+func OnConflict(fn func() error) error {
+	return retry.RetryOnConflict(wait.Backoff{
+		Duration: 1500 * time.Millisecond,
+		Factor:   1.5,
+		Jitter:   0,
+		Steps:    9999,
+		Cap:      30 * time.Second,
+	}, fn)
+}